refactor: leave pending records to the watcher and heal failed sends once per open
TransactionStatusWatcher already polls every pending record app-wide,
history page included, so the page-mounted sweep was a second reader of
the same records: it duplicated the chain request outside react-query's
dedup and could apply the same confirmation twice, defeating the
once-per-verdict guard. The sweep now only does the one job the watcher
does not — healing sends an older client wrongly failed — and checks
each candidate once per mount rather than on every records refetch.
